samples/requirements.txt
onnx
numpy
onnxruntime
onnxruntime-openvino
tritonclient
grpcio>=1.63.0,<1.68
numpy>=1.19.1
packaging>=14.1
protobuf>=5.26.1,<6.0dev
python-rapidjson>=0.9.1